pyproject.toml
[project]
name = "nonebot-plugin-diffsinger"
version = "0.1.11"
description = "用DiffSinger让bot唱歌"
authors = [{ name = "zhzhongshi", email = "zhz894524@163.com" }]
license = { text = "MIT" }
dependencies = [
    "nonebot2>=2.2.0,<3.0.0",
    "nonebot-adapter-onebot>=2.4.3",
    "httpx>=0.27.0",
    "librosa>=0.10.1",
]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/zhzhongshi/nonebot-plugin-diffsinger"
Repository = "https://github.com/zhzhongshi/nonebot-plugin-diffsinger"


[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
